+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Upcoming
+
+* Add hash based lookup for `DigestField` and `HMACField`.
+* Add `DigestField`, `HMACField`, `PGPPublicKeyAggregate`, `PGPSymmetricKeyAggregate`.
+
 ## v0.1.0
 
 * Add decryption through an aggregate class.